+++ src/com/dekarrin/file/png/CriticalChunk.java
@@ -19,7 +19,7 @@ public abstract class CriticalChunk extends Chunk {
 	 * @param crc
 	 * The cyclic redundancy check read from the chunk.
 	 */
-	public CriticalChunk(byte[] type, byte[] data, int crc) {
+	public CriticalChunk(byte[] type, byte[] data, long crc) {
 		super(type, data, crc);
 	}
 }
